api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from django.http import JsonResponse
from .views import UsuarioViewSet

router = DefaultRouter()
router.register(r'users', UsuarioViewSet, basename='usuario')

# Endpoint para Kubernetes probes
def health_check(request):
    return JsonResponse({"status": "ok"})

urlpatterns = [
    path('health/', health_check),         # ← nuevo endpoint
    path('', include(router.urls)),
]
